Move CoGrouped cascading logic into CascadingBackend
