+++ scalding-core/src/main/scala/com/twitter/scalding/typed/CoGrouped.scala
@@ -15,18 +15,7 @@ limitations under the License.
 */
 package com.twitter.scalding.typed
 
-import cascading.flow.FlowDef
-import cascading.tuple.{ Tuple => CTuple, Fields }
-import cascading.pipe.joiner.{ Joiner => CJoiner, JoinerClosure }
-import cascading.pipe.{ CoGroup, Pipe }
-
-import com.twitter.scalding._
-
-import scala.collection.JavaConverters._
-
-import com.twitter.scalding.serialization.Externalizer
-import com.twitter.scalding.TupleConverter.tuple2Converter
-import com.twitter.scalding.TupleSetter.tup2Setter
+import cascading.tuple.{ Tuple => CTuple }
 
 object CoGrouped {
   // distinct by mapped, but don't reorder if the list is unique
@@ -203,167 +192,4 @@ trait CoGrouped[K, +R] extends KeyedListLike[K, R, CoGrouped] with CoGroupable[K
 
   override def toTypedPipe: TypedPipe[(K, R)] =
     TypedPipe.CoGroupedPipe(this)
-
-  private[typed] def toKVPipe(flowDef: FlowDef, mode: Mode): Pipe = {
-    // Cascading handles the first item in join differently, we have to see if it is repeated
-    val firstCount = inputs.count(_ == inputs.head)
-
-    import Dsl._
-    import RichPipe.assignName
-
-    /*
-     * we only want key and value.
-     * Cascading requires you have the same number coming in as out.
-     * in the first case, we introduce (null0, null1), in the second
-     * we have (key1, value1), but they are then discarded:
-     */
-    def outFields(inCount: Int): Fields =
-      List("key", "value") ++ (0 until (2 * (inCount - 1))).map("null%d".format(_))
-
-    // Make this stable so the compiler does not make a closure
-    val ord = keyOrdering
-
-    val newPipe = Grouped.maybeBox[K, Any](ord, flowDef) { (tupset, ordKeyField) =>
-      if (firstCount == inputs.size) {
-        /**
-         * This is a self-join
-         * Cascading handles this by sending the data only once, spilling to disk if
-         * the groups don't fit in RAM, then doing the join on this one set of data.
-         * This is fundamentally different than the case where the first item is
-         * not repeated. That case is below
-         */
-        val NUM_OF_SELF_JOINS = firstCount - 1
-        new CoGroup(assignName(inputs.head.toPipe[(K, Any)](("key", "value"))(flowDef, mode,
-          tupset)),
-          ordKeyField,
-          NUM_OF_SELF_JOINS,
-          outFields(firstCount),
-          WrappedJoiner(new DistinctCoGroupJoiner(firstCount, Grouped.keyGetter(ord), joinFunction)))
-      } else if (firstCount == 1) {
-
-        def keyId(idx: Int): String = "key%d".format(idx)
-        /**
-         * As long as the first one appears only once, we can handle self joins on the others:
-         * Cascading does this by maybe spilling all the streams other than the first item.
-         * This is handled by a different CoGroup constructor than the above case.
-         */
-        def renamePipe(idx: Int, p: TypedPipe[(K, Any)]): Pipe =
-          p.toPipe[(K, Any)](List(keyId(idx), "value%d".format(idx)))(flowDef, mode,
-            tupset)
-
-        // This is tested for the properties we need (non-reordering)
-        val distincts = CoGrouped.distinctBy(inputs)(identity)
-        val dsize = distincts.size
-        val isize = inputs.size
-
-        def makeFields(id: Int): Fields = {
-          val comp = ordKeyField.getComparators.apply(0)
-          val fieldName = keyId(id)
-          val f = new Fields(fieldName)
-          f.setComparator(fieldName, comp)
-          f
-        }
-
-        val groupFields: Array[Fields] = (0 until dsize)
-          .map(makeFields)
-          .toArray
-
-        val pipes: Array[Pipe] = distincts
-          .zipWithIndex
-          .map { case (item, idx) => assignName(renamePipe(idx, item)) }
-          .toArray
-
-        val cjoiner = if (isize != dsize) {
-          // avoid capturing anything other than the mapping ints:
-          val mapping: Map[Int, Int] = inputs.zipWithIndex.map {
-            case (item, idx) =>
-              idx -> distincts.indexWhere(_ == item)
-          }.toMap
-
-          new CoGroupedJoiner(isize, Grouped.keyGetter(ord), joinFunction) {
-            val distinctSize = dsize
-            def distinctIndexOf(orig: Int) = mapping(orig)
-          }
-        } else {
-          new DistinctCoGroupJoiner(isize, Grouped.keyGetter(ord), joinFunction)
-        }
-
-        new CoGroup(pipes, groupFields, outFields(dsize), WrappedJoiner(cjoiner))
-      } else {
-        /**
-         * This is non-trivial to encode in the type system, so we throw this exception
-         * at the planning phase.
-         */
-        sys.error("Except for self joins, where you are joining something with only itself,\n" +
-          "left-most pipe can only appear once. Firsts: " +
-          inputs.collect { case x if x == inputs.head => x }.toString)
-      }
-    }
-    /*
-       * the CoGrouped only populates the first two fields, the second two
-       * are null. We then project out at the end of the method.
-       */
-    val pipeWithRedAndDescriptions = {
-      RichPipe.setReducers(newPipe, reducers.getOrElse(-1))
-      RichPipe.setPipeDescriptions(newPipe, descriptions)
-      newPipe.project('key, 'value)
-    }
-    pipeWithRedAndDescriptions
-  }
-}
-
-abstract class CoGroupedJoiner[K](inputSize: Int,
-  getter: TupleGetter[K],
-  @transient inJoinFunction: (K, Iterator[CTuple], Seq[Iterable[CTuple]]) => Iterator[Any]) extends CJoiner {
-
-  /**
-   * We have a test that should fail if Externalizer is not used here.
-   * you can test failure of that test by replacing Externalizer with Some
-   */
-  val joinFunction = Externalizer(inJoinFunction)
-  val distinctSize: Int
-  def distinctIndexOf(originalPos: Int): Int
-
-  // This never changes. Compute it once
-  protected val restIndices: IndexedSeq[Int] = (1 until inputSize).map { idx =>
-    val didx = distinctIndexOf(idx)
-    assert(didx > 0, "the left most can only be iterated once")
-    didx
-  }
-
-  override def getIterator(jc: JoinerClosure) = {
-    val iters = (0 until distinctSize).map { jc.getIterator(_).asScala.buffered }
-    // This use of `_.get` is safe, but difficult to prove in the types.
-    @SuppressWarnings(Array("org.wartremover.warts.OptionPartial"))
-    val keyTuple = iters
-      .collectFirst { case iter if iter.nonEmpty => iter.head }
-      .get // One of these must have a key
-    val key = getter.get(keyTuple, 0)
-
-    val leftMost = iters.head
-
-    def toIterable(didx: Int) =
-      new Iterable[CTuple] { def iterator = jc.getIterator(didx).asScala }
-
-    val rest = restIndices.map(toIterable(_))
-    joinFunction.get(key, leftMost, rest).map { rval =>
-      // There always has to be the same number of resulting fields as input
-      // or otherwise the flow planner will throw
-      val res = CTuple.size(distinctSize)
-      res.set(0, key)
-      res.set(1, rval)
-      res
-    }.asJava
-  }
-
-  override def numJoins = distinctSize - 1
-}
-
-// If all the input pipes are unique, this works:
-class DistinctCoGroupJoiner[K](count: Int,
-  getter: TupleGetter[K],
-  @transient joinF: (K, Iterator[CTuple], Seq[Iterable[CTuple]]) => Iterator[Any])
-  extends CoGroupedJoiner[K](count, getter, joinF) {
-  val distinctSize = count
-  def distinctIndexOf(idx: Int) = idx
 }
